+++ R/calc_ale.R
@@ -374,8 +374,19 @@ calc_ale <- function(
       # Generate the cumulative ALE y predictions.
       if (ixn_d == 1) {
         if (xd[[x_cols]]$x_type == 'numeric') {
-          # For 1D ALE, set origin effect for minimum numeric value to zero; there should be no other missing values.
+          # For 1D ALE, set origin effect for minimum numeric value to zero
           btit.local_eff_ray[it.cat, 1] <- 0
+
+          # There should be no other missing values except perhaps in the case of full-model bootstrapping
+          it.na_idx <- is.na(btit.local_eff_ray[it.cat, ])
+          if (any(it.na_idx)) {
+            if (is.null(.bins)) {
+              cli_abort('There should be no other missing values in 1D ALE at this point. Please submit a bug report.')
+            } else {
+              # Full-model bootstrapping might occasionally leave missing bins here, so just impute them with zero accumulation
+              btit.local_eff_ray[it.cat, it.na_idx] <- 0
+            }
+          }
         }
 
         # Accumulate the effects.